Separate models from JSON serialization specs
